requirements.txt
certifi==2025.1.31
gunicorn==20.1.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==2.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Pillow==12.1.0
python-dotenv==1.2.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==2.2.2
